Move `is_synched` from BCS to RPCClient
